+++ src/app/page.tsx
@@ -1,6 +1,9 @@
+import AuthButton from "../components/AuthButton";
+
 export default function Home() {
   return (
     <main>
+      <AuthButton />
       <div className="w-full px-8 py-12 md:py-20 flex flex-col items-center">
         <h1 className="text-center text-4xl md:text-6xl max-w-xl font-semibold">
           PaperHans